Separated helper methods into files
